Update README.md
This project analyzes the effectiveness of two different marketing strategies, A and B, using A/B testing techniques in Python. By conducting a hypothesis test on sales data from two groups of randomly selected customers, we determine which marketing strategy is more effective and provide a recommendation to the marketing firm.